+++ genwasym_runtime/include/wasm/controls.hpp
@@ -1,103 +1,54 @@
-
 #ifndef WASM_CONTROLS_HPP
 #define WASM_CONTROLS_HPP
 
-#include <cassert>
 #include <functional>
-
-#include <iostream>
 #include <memory>
 #include <variant>
 
 class MContRepr;
+
 struct MCont_t {
   std::shared_ptr<MContRepr> ptr;
-  MCont_t() : ptr(nullptr) {}
-  MCont_t(const MCont_t &p) : ptr(p.ptr) {}
-  MCont_t(std::shared_ptr<MContRepr> p) : ptr(p) {}
-  MCont_t(std::function<std::monostate(std::monostate)> haltK)
-      : ptr(std::make_shared<MContRepr>(haltK)) {
-    assert(haltK);
-  }
-  bool is_null() const { return ptr == nullptr; }
 
+  MCont_t();
+  MCont_t(const MCont_t &p);
+  MCont_t(std::shared_ptr<MContRepr> p);
+  MCont_t(std::function<std::monostate(std::monostate)> haltK);
+
+  bool is_null() const;
   std::monostate enter();
 };
+
 using Cont_t = std::function<std::monostate(std::monostate)>;
 
-static MCont_t CURRENT_MCONT;
+extern MCont_t CURRENT_MCONT;
 
-inline std::monostate updateCurrentMCont(MCont_t newMCont) {
-  CURRENT_MCONT = newMCont;
-  return std::monostate{};
-}
+std::monostate updateCurrentMCont(MCont_t newMCont);
 
 class MContRepr {
   friend std::monostate enterCC(std::monostate);
 
 public:
-  MContRepr(Cont_t cont, MCont_t mcont) : cont(cont), mcont(mcont) {}
+  MContRepr(Cont_t cont, MCont_t mcont);
+  MContRepr(std::function<std::monostate(std::monostate)> haltK);
 
-  MContRepr(std::function<std::monostate(std::monostate)> haltK)
-      : cont(haltK), mcont() {}
-
-  // MContRepr() : cont(nullptr), mcont() {}
-
-  std::monostate enter() {
-    // std::cout << "Entering MCont\n";
-    // std::cout << "Cont cont: " << (cont ? "valid" : "null") << "\n";
-    // std::cout << "MCont mcont: " << (mcont ? "valid" : "null") << "\n";
-
-    // This is necessary, because `this` may be deleted
-    // after next line. This copy is cheap because we always store a function
-    // pointer (non captured free variable lambda) in cont.
-    std::monostate (*func_ptr)(std::monostate) = nullptr;
-    {
-      auto cont = this->cont;
-      func_ptr = *cont.target<std::monostate (*)(std::monostate)>();
-    }
-
-    CURRENT_MCONT = mcont;
-
-    return func_ptr(std::monostate{});
-  }
+  std::monostate enter();
 
 private:
   Cont_t cont;
   MCont_t mcont;
 };
 
-inline MCont_t prependCont(Cont_t k, MCont_t mcont) {
-  return std::make_shared<MContRepr>(k, mcont);
-}
-
-inline std::monostate MCont_t::enter() { return ptr->enter(); }
+MCont_t prependCont(Cont_t k, MCont_t mcont);
 
 // Enter the current global MCont (CURRENT_MCONT)
-inline std::monostate enterCC(std::monostate) {
-  // std::cout << "Entering MCont\n";
-  // std::cout << "Cont cont: " << (cont ? "valid" : "null") << "\n";
-  // std::cout << "MCont mcont: " << (mcont ? "valid" : "null") << "\n";
-
-  // This is necessary, because `this` may be deleted
-  // after next line. This copy is cheap because we always store a function
-  // pointer (non captured free variable lambda) in cont.
-  std::monostate (*func_ptr)(std::monostate) = nullptr;
-  {
-    auto cont = CURRENT_MCONT.ptr->cont;
-    func_ptr = *cont.target<std::monostate (*)(std::monostate)>();
-  }
-
-  CURRENT_MCONT = CURRENT_MCONT.ptr->mcont;
-
-  __attribute__((musttail)) return func_ptr(std::monostate{});
-}
+std::monostate enterCC(std::monostate);
 
 struct Control {
   Cont_t cont;
   MCont_t mcont;
 
-  Control(Cont_t cont, MCont_t mcont) : cont(cont), mcont(mcont) {}
+  Control(Cont_t cont, MCont_t mcont);
 };
 
 using Func_t = std::function<std::monostate(std::monostate)>;
