+++ genwasym_runtime/include/wasm/z3_env.hpp
@@ -1,36 +1,24 @@
 #ifndef WASM_Z3_ENV_HPP
 #define WASM_Z3_ENV_HPP
+
 #include "z3++.h"
+#include <unordered_map>
 
 struct Z3Env {
   z3::context z3_ctx;
 
-  Z3Env() : z3_ctx() {}
+  Z3Env();
 };
 
-static Z3Env GLOBAL_Z3_ENV;
+extern Z3Env GLOBAL_Z3_ENV;
 
-inline z3::context &global_z3_ctx() { return GLOBAL_Z3_ENV.z3_ctx; }
+z3::context &global_z3_ctx();
 
 // A map from z3 expression id to their ast size
-static std::unordered_map<unsigned, int> Z3ExprSizeMap;
+extern std::unordered_map<unsigned, int> Z3ExprSizeMap;
 
-inline int get_z3_fp_sort_size(const z3::sort &s) {
-  assert(s.is_fpa());
-  return s.fpa_ebits() + s.fpa_sbits();
-}
+int get_z3_fp_sort_size(const z3::sort &s);
 
-static int get_z3_expr_size(const z3::expr &e) {
-  unsigned id = e.id();
-  if (Z3ExprSizeMap.find(id) != Z3ExprSizeMap.end()) {
-    return Z3ExprSizeMap[id];
-  }
-  unsigned count = 1; // count self
-  for (unsigned i = 0; i < e.num_args(); i++) {
-    count += get_z3_expr_size(e.arg(i));
-  }
-  Z3ExprSizeMap[id] = count;
-  return count;
-}
+int get_z3_expr_size(const z3::expr &e);
 
 #endif // WASM_Z3_ENV_HPP
